// src/models/player-model.ts
export interface playerModel {
    id: number;
    name: string;
    club: string;
    nationality: string;
    position: string;
    statistics: {
        overall: number;
        pace: number;
        shooting: number;
        passing: number;
        dribbling: number;
        defending: number;
        physical: number;
    };
};